+++ locales/en/index.yml
@@ -319,6 +319,9 @@ authentication:
       carousel:
         label: IO presentation
         hint: Carousel that contains an introduction guide to IO functionalities. Select to know the content.
+    session_expired:
+      title: Your session is expired
+      body: For security reasons, after 30 days from the last access, we ask you to log in again with SPID or CIE to use IO.
     cie_unsupported:
       title: Login with CIE unavailable
       body: !include cie/cieNotSupported.md
