+++ ts/features/bonusVacanze/components/loadingErrorScreen/LoadingErrorComponent.tsx
@@ -5,7 +5,7 @@ import { useEffect } from "react";
 import { ActivityIndicator, SafeAreaView } from "react-native";
 import GenericErrorComponent from "../../../../components/screens/GenericErrorComponent";
 import { setAccessibilityFocus } from "../../../../utils/accessibility";
-import { InfoScreenComponent } from "../infoScreen/InfoScreenComponent";
+import { InfoScreenComponent } from "../../../../components/infoScreen/InfoScreenComponent";
 import { bonusVacanzeStyle } from "../Styles";
 
 export type LoadingErrorProps = {
@@ -57,15 +57,10 @@ const delay = 100 as Millisecond;
  * @param props
  * @constructor
  */
-export const LoadingErrorComponent: React.FunctionComponent<
-  LoadingErrorProps
-> = props => {
-  useEffect(
-    () => {
-      setAccessibilityFocus(props.isLoading ? loadingRef : errorRef, delay);
-    },
-    [props.isLoading]
-  );
+export const LoadingErrorComponent: React.FunctionComponent<LoadingErrorProps> = props => {
+  useEffect(() => {
+    setAccessibilityFocus(props.isLoading ? loadingRef : errorRef, delay);
+  }, [props.isLoading]);
 
   return (
     <SafeAreaView style={bonusVacanzeStyle.flex}>
