+++ ts/features/bonusVacanze/screens/activation/EligibilityExpired.tsx
@@ -9,8 +9,8 @@ import {
   confirmButtonProps
 } from "../../components/buttons/ButtonConfigurations";
 import { useHardwareBackButton } from "../../components/hooks/useHardwareBackButton";
-import { renderInfoRasterImage } from "../../components/infoScreen/imageRendering";
-import { InfoScreenComponent } from "../../components/infoScreen/InfoScreenComponent";
+import { renderInfoRasterImage } from "../../../../components/infoScreen/imageRendering";
+import { InfoScreenComponent } from "../../../../components/infoScreen/InfoScreenComponent";
 import { bonusVacanzeStyle } from "../../components/Styles";
 import {
   cancelBonusVacanzeRequest,
@@ -66,7 +66,4 @@ const mapDispatchToProps = (dispatch: Dispatch) => ({
   onRestart: () => dispatch(checkBonusVacanzeEligibility.request())
 });
 
-export default connect(
-  null,
-  mapDispatchToProps
-)(EligibilityExpired);
+export default connect(null, mapDispatchToProps)(EligibilityExpired);
