+++ ts/features/bonusVacanze/screens/eligibility/isee/IseeNotAvailableScreen.tsx
@@ -5,7 +5,7 @@ import { Dispatch } from "redux";
 import { openLink } from "../../../../../components/ui/Markdown/handlers/link";
 
 import I18n from "../../../../../i18n";
-import { InfoScreenStyle } from "../../../components/infoScreen/InfoScreenComponent";
+import { InfoScreenStyle } from "../../../../../components/infoScreen/InfoScreenComponent";
 import { cancelBonusVacanzeRequest } from "../../../store/actions/bonusVacanze";
 import { BaseIseeErrorComponent } from "./BaseIseeErrorComponent";
 
@@ -63,7 +63,4 @@ const mapDispatchToProps = (dispatch: Dispatch) => ({
   onCancel: () => dispatch(cancelBonusVacanzeRequest())
 });
 
-export default connect(
-  null,
-  mapDispatchToProps
-)(IseeNotAvailableScreen);
+export default connect(null, mapDispatchToProps)(IseeNotAvailableScreen);
